test: fold a11y axe scans into main component tests
Each *.a11y.test.tsx was an isolated file with a single axe scan plus
(in two cases) substantive keyboard/aria behavior tests. The mount
setup was duplicated from the matching *.test.tsx, and the axe scan
itself often ran a second time in the main file.

- Fretboard / NoteGrid / StepperControl / ToggleBar: replaced the
  single-axe-scan file with one `it("has no a11y violations")`
  inside the existing describe (ToggleBar parameterized over both
  variants).
- CircleOfFifths: ported 10 keyboard-navigation tests into a
  "Keyboard navigation & a11y" describe and collapsed paired arrow
  keys / Enter+Space into `it.each` blocks.
- FretboardSVG: ported the meaningful subset of FretboardNoteLayer's
  a11y suite (octave correctness, tabIndex toggling on onNoteClick,
  Enter/Space activation) into a "note layer a11y contract" describe;
  dropped the duplicate axe scan (FretboardSVG.test.tsx already has
  one) and the duplicate click test.

Test count: 1941 -> 1937 (-4 via collapsed it.each and removed dupes).
Test files: 112 -> 106. All passing.
